+++ src/fromir.rs
@@ -1,10 +1,13 @@
-use crate::common::{Variant};
+use crate::common::{Variant, Colour};
 use half::f16;
 use std::convert::TryInto;
 use std::str::{from_utf8, from_utf8_unchecked};
 use std::net::{SocketAddrV4, SocketAddrV6};
 use std::io::Write;
 use std::borrow::Cow;
+
+use ansi_term::{ANSIGenericString, Style};
+use std::env::var;
 
 pub enum Endianness {
     Default,
@@ -18,7 +21,7 @@ pub trait FromIR {
     fn variants(ir: & [u8]) -> Option<Vec<Variant>>;
 
     ///Used to convert an IR to a String. NOTE: this function does NOT check to make sure that the IR can be converted, and will panic if the conversion fails
-    fn encode(ir: & [u8], variant: Variant) -> String;
+    fn encode(ir: & [u8], variant: Variant) -> ANSIGenericString<str>;
 
     fn endianness() -> Endianness;
 }
@@ -32,7 +35,7 @@ impl FromIR for crate::common::Base2_16 {
              Variant("Base 16")])
     }
 
-    fn encode(ir: & [u8], variant: Variant) -> String {
+    fn encode(ir: & [u8], variant: Variant) -> ANSIGenericString<str> {
         let base = Self::get_base(&variant);
 
         let mut base_n_list = convert_base::Convert::new(256, base).convert::<u8, u8>(ir.as_ref());
@@ -43,7 +46,7 @@ impl FromIR for crate::common::Base2_16 {
             string.insert(0, Self::num_to_ascii(*byte, true) as char)
         }
 
-        string
+        Style::default().paint(string)
     }
 
     fn endianness() -> Endianness {
@@ -64,20 +67,20 @@ impl FromIR for crate::common::FixedFloat {
 
     }
 
-    fn encode(ir: & [u8], variant: Variant) -> String {
+    fn encode(ir: & [u8], variant: Variant) -> ANSIGenericString<str> {
 
         if variant.0.len() == 24 {
             if &variant.0[7..] == "mantissa/exponent" {
-                return Self::to_mes_string(ir);
+                return Style::default().paint(Self::to_mes_string(ir));
             }
         }
 
-        match &variant.0[0..6] {
+        Style::default().paint(match &variant.0[0..6] {
             "16-bit" => f16::from_le_bytes(ir.as_ref().try_into().unwrap()).to_f64().to_string(),
             "32-bit" => (f32::from_le_bytes(ir.as_ref().try_into().unwrap()) as f64).to_string(),
             "64-bit" => f64::from_le_bytes(ir.as_ref().try_into().unwrap()).to_string(),
             _ => panic!("Invalid variant in FromIT FixedFloat")
-        }
+        })
     }
 
     fn endianness() -> Endianness {
@@ -99,8 +102,8 @@ impl FromIR for crate::common::FixedInt {
         }
     }
 
-    fn encode(ir: & [u8], variant: Variant) -> String {
-        match variant.0 {
+    fn encode(ir: & [u8], variant: Variant) -> ANSIGenericString<str> {
+        Style::default().paint(match variant.0 {
             "u8" => u8::from_le_bytes(ir.as_ref().try_into().unwrap()).to_string(),
             "i8" => i8::from_le_bytes(ir.as_ref().try_into().unwrap()).to_string(),
             "u16" => u16::from_le_bytes(ir.as_ref().try_into().unwrap()).to_string(),
@@ -112,7 +115,7 @@ impl FromIR for crate::common::FixedInt {
             "u128" => u128::from_le_bytes(ir.as_ref().try_into().unwrap()).to_string(),
             "i128" => i128::from_le_bytes(ir.as_ref().try_into().unwrap()).to_string(),
             _ => panic!("Invalid variant in FromIT FixedInt")
-        }
+        })
     }
 
     fn endianness() -> Endianness {
@@ -142,7 +145,7 @@ impl FromIR for crate::common::DateTime {
         }
     }
 
-    fn encode(ir: & [u8], variant: Variant) -> String {
+    fn encode(ir: & [u8], variant: Variant) -> ANSIGenericString<str> {
 
         let timestamp = match &(variant.0)[0..2] {
             "64" => i64::from_le_bytes(ir.as_ref().try_into().unwrap()),
@@ -152,12 +155,12 @@ impl FromIR for crate::common::DateTime {
 
         let datetime = chrono::DateTime::<chrono::Utc>::from_utc(chrono::NaiveDateTime::from_timestamp(timestamp, 0), chrono::Utc);
 
-        match &(variant.0)[7..14] {
+        Style::default().paint(match &(variant.0)[7..14] {
             "rfc2822" => datetime.to_rfc2822(),
             "rfc3339" => datetime.to_rfc3339(),
             "custom " => datetime.format(&(variant.0)[14..]).to_string(),
             _ => panic!("Invalid variant (format) in FromIT DateTime")
-        }
+        })
 
     }
 
@@ -169,17 +172,41 @@ impl FromIR for crate::common::DateTime {
 impl FromIR for crate::common::Unicode8 {
     fn variants(ir: &[u8]) -> Option<Vec<Variant>> {
         if from_utf8(ir).is_ok() {
-            Some(vec![Variant("")])
+            for character in unsafe {from_utf8_unchecked(ir).chars()} {
+                if unicode_names2::name(character).is_none() {
+                    return Some(vec![Variant("Literal String")])
+                }
+            }
+
+            Some(vec![Variant("Literal String"), Variant("Unicode Names")])
+
         } else {
             None
         }
     }
 
-    fn encode(ir: &[u8], variant: Variant) -> String {
-        if variant.0 == "" {
+    fn encode(ir: &[u8], variant: Variant) -> ANSIGenericString<str> {
+
+        if variant.0 == "Literal String" {
             unsafe {
-                String::from(from_utf8_unchecked(ir))
+                Style::default().paint(String::from(from_utf8_unchecked(ir)))
             }
+        } else if variant.0 == "Unicode Names" {
+            let mut string = String::from("[");
+
+            let unicode = unsafe {
+                from_utf8_unchecked(ir)
+            };
+
+            for character in unicode.chars() {
+
+                string.push_str(&unicode_names2::name(character).unwrap().to_string());
+                string.push_str(" '");
+                string.push(character);
+                string.push_str("', ")
+            }
+
+            Style::default().paint(string)
         } else {
             panic!("Invalid variant in FromIR Unicode8");
         }
@@ -201,11 +228,11 @@ impl FromIR for crate::common::IpV4 {
         }
     }
 
-    fn encode(ir: &[u8], variant: Variant) -> String {
+    fn encode(ir: &[u8], variant: Variant) -> ANSIGenericString<str> {
 
         let ip = std::net::Ipv4Addr::from(u32::from_le_bytes((&ir[0..4]).try_into().unwrap()));
 
-        match variant.0 {
+        Style::default().paint(match variant.0 {
             "with port" => {
                 let port = u16::from_le_bytes((&ir[4..6]).try_into().unwrap());
 
@@ -216,7 +243,7 @@ impl FromIR for crate::common::IpV4 {
                 ip.to_string()
             }
             _ => panic!("Invalid variant in FromIT IpV4"),
-        }
+        })
     }
 
     fn endianness() -> Endianness {
@@ -235,11 +262,11 @@ impl FromIR for crate::common::IpV6 {
         }
     }
 
-    fn encode(ir: &[u8], variant: Variant) -> String {
+    fn encode(ir: &[u8], variant: Variant) -> ANSIGenericString<str> {
 
         let ip = std::net::Ipv6Addr::from(u128::from_le_bytes((&ir[0..16]).try_into().unwrap()));
 
-        match variant.0 {
+        Style::default().paint(match variant.0 {
             "with port" => {
                 let port = u16::from_le_bytes((&ir[16..18]).try_into().unwrap());
 
@@ -250,7 +277,7 @@ impl FromIR for crate::common::IpV6 {
                 ip.to_string()
             }
             _ => panic!("Invalid variant in FromIT IpV4"),
-        }
+        })
     }
 
     fn endianness() -> Endianness {
@@ -263,9 +290,9 @@ impl FromIR for crate::common::Base91 {
         Some(vec![Variant("")])
     }
 
-    fn encode(ir: &[u8], _variant: Variant) -> String {
+    fn encode(ir: &[u8], _variant: Variant) -> ANSIGenericString<str> {
         unsafe {
-            String::from_utf8_unchecked(base91::slice_encode(ir))
+            Style::default().paint(String::from_utf8_unchecked(base91::slice_encode(ir)))
         }
     }
 
@@ -279,12 +306,12 @@ impl FromIR for crate::common::Base85 {
         Some(vec![Variant("z85"), Variant("ascii85")])
     }
 
-    fn encode(ir: &[u8], variant: Variant) -> String {
-        match variant.0 {
+    fn encode(ir: &[u8], variant: Variant) -> ANSIGenericString<str> {
+        Style::default().paint(match variant.0 {
             "z85" => z85::encode(ir),
              "ascii85" => ascii85::encode(ir),
             _ => panic!("Invalid variant in FromIR Base85")
-        }
+        })
     }
 
     fn endianness() -> Endianness {
@@ -304,8 +331,8 @@ impl FromIR for crate::common::Base64 {
              Variant("URL-safe no padding"),])
     }
 
-    fn encode(ir: &[u8], variant: Variant) -> String {
-        match variant.0 {
+    fn encode(ir: &[u8], variant: Variant) -> ANSIGenericString<str> {
+        Style::default().paint(match variant.0 {
             "Bcrypt" => base64::encode_config(ir, base64::BCRYPT),
             "BinHex" => base64::encode_config(ir, base64::BINHEX),
             "crypt" => base64::encode_config(ir, base64::CRYPT),
@@ -315,7 +342,7 @@ impl FromIR for crate::common::Base64 {
             "URL-safe" => base64::encode_config(ir, base64::STANDARD_NO_PAD),
             "URL-safe no padding" => base64::encode_config(ir, base64::URL_SAFE_NO_PAD),
             _ => panic!("Invalid variant in FromIr Base64"),
-        }
+        })
     }
 
     fn endianness() -> Endianness {
@@ -328,38 +355,8 @@ impl FromIR for crate::common::ByteList {
         Some(vec![Variant("")])
     }
 
-    fn encode(ir: &[u8], _variant: Variant) -> String {
-        format!("({} byte(s)) {:?}", ir.len(), ir)
-    }
-
-    fn endianness() -> Endianness {
-        Endianness::Default
-    }
-}
-
-impl FromIR for crate::common::Hash {
-    fn variants(_ir: &[u8]) -> Option<Vec<Variant>> {
-        Some(vec![Variant("md5"), Variant("sha1"), Variant("sha256"), Variant("sha512")])
-    }
-
-    fn encode(ir: &[u8], variant: Variant) -> String {
-        let algorithm = match variant.0 {
-            "md5" => crypto_hash::Algorithm::MD5,
-            "sha1" => crypto_hash::Algorithm::SHA1,
-            "sha256" => crypto_hash::Algorithm::SHA256,
-            "sha512" => crypto_hash::Algorithm::SHA512,
-            _ => panic!("Invalid variant FromIT Hash")
-        };
-
-        let mut hasher = crypto_hash::Hasher::new(algorithm);
-
-        hasher.write_all(ir).unwrap();
-
-        let mut hash = hasher.finish();
-
-        hash.reverse();
-
-        crate::common::Base2_16::encode(&hash, Variant("Base 16"))
+    fn encode(ir: &[u8], _variant: Variant) -> ANSIGenericString<str> {
+        Style::default().paint(format!("({} byte(s)) {:?}", ir.len(), ir))
     }
 
     fn endianness() -> Endianness {
@@ -376,9 +373,9 @@ impl FromIR for crate::common::UUID {
         }
     }
 
-    fn encode(ir: &[u8], variant: Variant) -> String {
+    fn encode(ir: &[u8], variant: Variant) -> ANSIGenericString<str> {
         if variant.0 == "" {
-            uuid::Uuid::from_slice(ir).unwrap().to_string()
+            Style::default().paint(uuid::Uuid::from_slice(ir).unwrap().to_string())
         } else {
             panic!("Invalid variant in FromIT uuid")
         }
@@ -394,9 +391,9 @@ impl FromIR for crate::common::EscapedString {
         Some(vec![Variant("")])
     }
 
-    fn encode(ir: &[u8], variant: Variant) -> String {
+    fn encode(ir: &[u8], variant: Variant) -> ANSIGenericString<str> {
         if variant.0 == "" {
-            crate::escape::EscapeSequence::encode(ir)
+            Style::default().paint(crate::escape::EscapeSequence::encode(ir))
         } else {
             panic!("Invalid variant in FromIR EscapedString");
         }
@@ -412,9 +409,9 @@ impl FromIR for crate::common::UrlEncode {
         Some(vec![Variant("")])
     }
 
-    fn encode(ir: &[u8], variant: Variant) -> String {
+    fn encode(ir: &[u8], variant: Variant) -> ANSIGenericString<str> {
         if variant.0 == "" {
-            urlencoding::encode_binary(ir).to_string()
+            Style::default().paint(urlencoding::encode_binary(ir).to_string())
         } else {
             panic!("Invalid variant in FromIR UrlEncode");
         }
@@ -434,7 +431,7 @@ impl FromIR for crate::common::UrlDecode {
         }
     }
 
-    fn encode(ir: &[u8], variant: Variant) -> String {
+    fn encode(ir: &[u8], variant: Variant) -> ANSIGenericString<str> {
 
         let string = if variant.0 == "RFC 3986" {
             let string = String::from_utf8_lossy(ir);
@@ -444,7 +441,7 @@ impl FromIR for crate::common::UrlDecode {
         };
 
         if variant.0 == "RFC 3986" || variant.0 == "Legacy" {
-            urlencoding::decode(&string).unwrap()
+            Style::default().paint(urlencoding::decode(&string).unwrap())
         } else {
             panic!("Invalid variant in FromIR UrlDecode");
         }
@@ -455,15 +452,90 @@ impl FromIR for crate::common::UrlDecode {
     }
 }
 
+impl FromIR for crate::common::Colour {
+    fn variants(ir: &[u8]) -> Option<Vec<Variant>> {
+        let variants = match ir.len() {
+            1 => vec![Variant("8-bit color"), Variant("8-bit greyscale"), Variant("8-bit terminal")],
+            2 => vec![Variant("16-bit color")],
+            3 => vec![Variant("24-bit color"), Variant("24-bit rgb"), Variant("24-bit hsl")],
+            4 => vec![Variant("32-bit color")],
+            _ => return None
+        };
+
+        Some(variants)
+    }
+
+    fn encode(ir: &[u8], variant: Variant) -> ANSIGenericString<str> {
+        use ansi_term::Color;
+
+        let colour = match variant.0 {
+            "8-bit color" => {
+                Colour::_8_to_24(*ir.get(0).unwrap())
+            },
+            "8-bit greyscale" => { let byte = *ir.get(0).unwrap(); Color::RGB(byte, byte, byte) },
+            "8-bit terminal" => { Color::Fixed(*ir.get(0).unwrap()) },
+            "16-bit color" => {Colour::_16_to_24(u16::from_le_bytes(ir.try_into().unwrap()))},
+            "24-bit color" | "24-bit rgb" | "24-bit hsl" => {
+                let red = *ir.get(2).unwrap();
+                let green = *ir.get(1).unwrap();
+                let blue = *ir.get(0).unwrap();
+
+                Color::RGB(red, green, blue)
+            },
+            "32-bit color" => {
+                let red = *ir.get(2).unwrap();
+                let green = *ir.get(1).unwrap();
+                let blue = *ir.get(0).unwrap();
+                Color::RGB(red, green, blue)
+            },
+            _ => panic!("Invalid variant in FromIR Colour")
+        };
+
+        match variant.0 {
+            "24-bit rgb" => {
+                if let Color::RGB(r, g, b) = colour {
+                    Style::default().paint(format!("rgb({}, {}, {})", r, g, b))
+                } else {
+                    panic!("Invalid variant in FromIR colour");
+                }
+
+            },
+            "24-bit hsl" => {
+                if let Color::RGB(r, g, b) = colour {
+
+                    let slice = [b, g, r];
+
+                    let hsl_value = hsl::HSL::from_rgb(&slice);
+
+                    Style::default().paint(format!("hsl({}, {}, {})", hsl_value.h, hsl_value.s, hsl_value.l))
+                } else {
+                    panic!("Invalid variant in FromIR colour");
+                }
+            },
+            _ => Style::default().fg(colour).paint("⬛"),
+        }
+
+
+    }
+
+    fn endianness() -> Endianness {
+        Endianness::Dual
+    }
+}
+
 /*
 
-impl FromIT for crate::common:: {
+impl FromIR for crate::common:: {
     fn variants(ir: &[u8]) -> Option<Vec<Variant>> {
         unimplemented!()
     }
 
-    fn encode(ir: &[u8], variant: Variant) -> String {
+    fn encode(ir: &[u8], variant: Variant) -> ANSIGenericString<str> {
         unimplemented!()
+    }
+
+    fn endianness() -> Endianness {
+        Endianness::Default
     }
 }