+++ changelog.md
@@ -6,10 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [Unreleased]
 ### To Do
 - ToIR implementations:
-  - Html colours (8, 16, 24 and 32-bit)
   - File path (load bytes from file, implement as a command line option)  
 - FromIR implementations:
-  - Colours (8, 16, 24 and 32-bit)
   - Disassembly (16, 32 and 64-bit)
   - File path (save bytes from file, implement as a command line option, with optional prompt to confirm the save)
 - Add command line options for
@@ -20,12 +18,31 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Unfinished Ideas
 - None
 
+## [0.1.4] - 2021-06-26
+### Added
+- Unicode names which allows the user to 
+  - Convert a unicode name into a character
+  - Convert a string into a list of unicode names
+- Added `common::Colour` to represent colours of various bit depths. Supplies some functions to convert to 24-bit colour  
+- `FromIR` implementation for colours supporting
+  - 8, 16, 24 and 32-bit colours
+  - 8-bit colour, greyscale and terminal formats  
+  - RGB and HSL values for 24-bit
+- `ToIR` implementation of colours supporting html colour format (#RRGGBB)
+
+### Changed
+- `FromIR::encode` now returns `ANSIGenericString<str>` which allows styling of terminal strings
+- Showing dual endianness now only colours the brackets magenta to allow styled text
+
 ## [0.1.3] - 2021-06-12
 ### Added
 - Mantissa/Exponent form for `FixedFloat` added
 - Endianness
   - `ToIr`: Always store data as little endian
   - `FromIr`: Two options, display the default endianness (chosen by `FromIr::encode`) or both little and big if no default can be chosen (primitive integers and float for example)
+  - Endianness is implemented by
+    - Adding a function to `FromIR` that identifies if a type has default or dual endianness
+    - The main control flow by displaying a variants on their own (default) or two values for each variant, one little and another big endian (dual)
 
 ### Fixed
 - Escape sequence code now accepts \fff as a valid python sequence as well as a C sequence